+++ test/for_collection_test.rb
@@ -74,7 +74,7 @@ class ForCollectionTest < MiniTest::Spec
 end
 
 
-class ImplicitCollectionTest < MiniTest::Spec
+class RepresentTest < MiniTest::Spec
   let (:songs) { [song, Song.new("Can't Take Them All")] }
   let (:song) { Song.new("Days Go By") }
 
@@ -84,7 +84,7 @@ class ImplicitCollectionTest < MiniTest::Spec
     # :xml  => [Representable::XML,  out="<a><song></song><song></song></a>", out]
   ) do |format, mod, output, input|
 
-    # collection is automatically detected.
+    # Representer.represents detects collection.
     describe "Module#to_/from_#{format}" do
       let (:format) { format }
 
@@ -102,21 +102,22 @@ class ImplicitCollectionTest < MiniTest::Spec
       it { parse(representer.represent([]), input).must_equal songs }
     end
 
+    # Decorator.represents detects collection.
+    describe "Decorator#to_/from_#{format}" do
+      let (:format) { format }
+      let (:representer) {
+        Class.new(Representable::Decorator) do
+          include mod
+          extend Representable::Represent
+          property :name
 
-    # describe "Decorator::for_collection [#{format}]" do
-    #   let (:format) { format }
-    #   let (:representer) {
-    #     Class.new(Representable::Decorator) do
-    #       include mod
-    #       property :name
+          collection_representer :class => Song
+        end
+      }
 
-    #       collection_representer :class => Song
-    #     end
-    #   }
-
-    #   it { render(representer.for_collection.new(songs)).must_equal_document output }
-    #   it { parse(representer.for_collection.new([]), input).must_equal songs }
-    # end
+      it { render(representer.represent(songs)).must_equal_document output }
+      it { parse(representer.represent([]), input).must_equal songs }
+    end
   end
 
 
@@ -126,7 +127,7 @@ class ImplicitCollectionTest < MiniTest::Spec
     # :xml  => [Representable::XML,  out="<a><song></song><song></song></a>", out]
   ) do |format, mod, output, input|
 
-    # singular is automatically detected.
+    # Representer.represents detects singular.
     describe "Module#to_/from_#{format}" do
       let (:format) { format }
 
@@ -145,19 +146,21 @@ class ImplicitCollectionTest < MiniTest::Spec
     end
 
 
-    # describe "Decorator::for_collection [#{format}]" do
-    #   let (:format) { format }
-    #   let (:representer) {
-    #     Class.new(Representable::Decorator) do
-    #       include mod
-    #       property :name
+    # Decorator.represents detects singular.
+    describe "Decorator#to_/from_#{format}" do
+      let (:format) { format }
+      let (:representer) {
+        Class.new(Representable::Decorator) do
+          include mod
+          extend Representable::Represent
+          property :name
 
-    #       collection_representer :class => Song
-    #     end
-    #   }
+          collection_representer :class => Song
+        end
+      }
 
-    #   it { render(representer.for_collection.new(songs)).must_equal_document output }
-    #   it { parse(representer.for_collection.new([]), input).must_equal songs }
-    # end
+      it { render(representer.represent(song)).must_equal_document output }
+      it { parse(representer.represent(Song.new), input).must_equal song }
+    end
   end
 end